Linked document type selector and client detail screens
Now is posible to navigate between the document type selector screen and
the client detail screen. Clicking "Agregar" button will take the user
to the document selector screen. The back button in both screens was
also enabled, so. when the user click the back button of the detail
screen the main screen will be show to the user and when the user clicks
the back button of the document type selector screen the detail screen
will be displayed.